Move TaskQueue to a separate file
